pyproject.toml
[project]
name = "crib-daq-mgmt"
version = "0.1.0"
description = "CRIB DAQ management scripts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gspread>=6.1.4",
    "pytest>=8.3.4",
    "pyyaml>=6.0.2",
    "sockets>=1.0.0",
]
